test: ActiveBids
