+++ src/schema/v2/homeView/__tests__/HomeViewSection.test.ts
@@ -1414,81 +1414,6 @@ describe("HomeViewSection", () => {
     })
   })
 
-  describe("Your Active Bids", () => {
-    it("returns correct data", async () => {
-      const query = gql`
-        {
-          homeView {
-            section(id: "home-view-section-active-bids") {
-              __typename
-
-              component {
-                title
-              }
-
-              ... on HomeViewSectionArtworks {
-                artworksConnection(first: 2) {
-                  edges {
-                    node {
-                      title
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      `
-
-      const lots = [
-        {
-          sale_artwork: {
-            artwork: {
-              title: "Artwork 1",
-            },
-          },
-        },
-        {
-          sale_artwork: {
-            artwork: {
-              title: "Artwork 2",
-            },
-          },
-        },
-      ]
-
-      const context = {
-        accessToken: "424242",
-        lotStandingLoader: jest.fn().mockResolvedValue(lots),
-      }
-
-      const { homeView } = await runQuery(query, context)
-
-      expect(homeView.section).toMatchInlineSnapshot(`
-        Object {
-          "__typename": "HomeViewSectionArtworks",
-          "artworksConnection": Object {
-            "edges": Array [
-              Object {
-                "node": Object {
-                  "title": "Artwork 1",
-                },
-              },
-              Object {
-                "node": Object {
-                  "title": "Artwork 2",
-                },
-              },
-            ],
-          },
-          "component": Object {
-            "title": "Your Active Bids",
-          },
-        }
-      `)
-    })
-  })
-
   describe("HomeViewSectionTasks", () => {
     beforeAll(() => {
       mockIsFeatureFlagEnabled.mockImplementation((flag: string) => {
